Move mockdb into separate shareable location
